requirements.txt
google-cloud-aiplatform
vertexai
shapely<2.0.0
PyMuPDF
google-cloud-storage
reportlab

